==> 5_COVIDvaccination_data_preparation.ipynb <==
#package loading 
import sklearn
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from lifelines import CoxPHFitter
from dateutil.relativedelta import relativedelta
from scipy.stats import percentileofscore
import umap
from tqdm import tqdm
import cudf
import pickle
import sys

from utility import *
from preprocess import *

import warnings
warnings.filterwarnings("ignore")

*The tables below may have information of COVID-vaccine*

#death_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/death_cancer.pkl','rb'))
#diagnosis_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/diagnosis_cancer.pkl','rb'))
#encounter_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/encounter_cancer.pkl','rb'))
#lab_result_cm_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/lab_result_cm_cancer.pkl','rb'))
med_admin_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/med_admin_cancer.pkl','rb'))
obs_gen_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/obs_gen_cancer.pkl','rb'))
procedure_cancer=pickle.load( open('/data/datasets/changyuyin/2_COVIDwithCancer/procedure_cancer.pkl','rb'))
#demographics_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/demographics_cancer.pkl','rb'))
dispensing_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/dispensing_cancer.pkl','rb'))
immunization_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/immunization_cancer.pkl','rb'))
#lds_address_history_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/lds_address_history_cancer.pkl','rb'))
#obs_CLIN_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/obs_CLIN_cancer.pkl','rb'))
prescribing_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/prescribing_cancer.pkl','rb'))
#vital_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/vital_cancer.pkl','rb'))

# Define codes representing COVID-19 vaccination

cpt_codes = tuple(set(['91300','0004A','91300','0001A','91300','0002A','91300',
                       '0003A','91301','0011A','91301','0012A','91301','0013A',
                       '91302','0021A','91302','0022A','91303','0034A','91303',
                       '0031A','91304','0044A','91304','0041A','91304','0042A',
                       '91305','0054A','91305','0051A','91305','0052A','91305',
                       '0053A','91306','0064A','91307','0074A','91307','0071A',
                       '91307','0072A','91307','0073A','91308','0081A','91308',
                       '0082A','91308','0083A','91309','0094A','91309','0091A',
                       '91309','0092A','91309','0093A','91310','0104A','91311',
                       '0111A','91311','0112A','91311','0113A','91312','0124A',
                       '91313','0134A','91314','0144A','91315','0154A','91316',
                       '0164A','91317','0173A','91303','91301','91306','91301',
                       '91309','91311','91313','91314','91316','91300','91305',
                       '91305','91307','91308','91312','91315','91317','91304',
                       '91310']))
print(len(cpt_codes))

cvx_codes = tuple(set(['212','207','207','207','221','228','229','230','208','217',
                   '217','218','219','300','301','302','211','225','213']))
print(len(cvx_codes))

rxnorm_codes = tuple(set(['2623378','2623382','2610328','2610319','2610347',
                      '2606078','2603740','2598700','2593847','2592364',
                      '2587906','2587899','2587898','2587313','2587312',
                      '2583743','2559909','2550903','2479835','2479154',
                      '2479153','2479150','2479149','2479159','2479158',
                      '2479156','2479155','2477902','2470234','2468235',
                      '2465246','2465248','2465253','2465255','2397257',
                      '2397251','2370236','2370234','2205474','2205472',
                      '2047247','2047241','2463118','2395502','2284960',
                      '2395504','2367758']))
print(len(rxnorm_codes))

ndc_codes = tuple(set(['5926710001','5926710001','5926710001','5926710001',
                   '8077727310','8077727310','8077727310','0310122210',
                   '0310122210','5967658005','5967658005','8063110001',
                   '8063110001','8063110001','5926710251','5926710251',
                   '5926710251','5926710251','8077727310','5926710551',
                   '5926710551','5926710551','5926710551','5926700781',
                   '5926700784','5926700784','8077727505','8077727505',
                   '8077727505','8077727505','4928161820','8077727905',
                   '8077727905','8077727905','5926703041','8077728205',
                   '8077728205','5926705651','8077728302','5926706091',
                   '59267100001','59267100001','59267100001','59267100001',
                   '80777027310','80777027310','80777027310','00310122210',
                   '00310122210','59676058005','59676058005','80631100001',
                   '80631100001','80631100001','59267102501','59267102501',
                   '59267102501','59267102501','80777027310','59267105501',
                   '59267105501','59267105501','59267105501','59267007801',
                   '59267007804','59267007804','80777027505','80777027505',
                   '80777027505','80777027505','49281061820','80777027905',
                   '80777027905','80777027905','59267030401','80777028205',
                   '80777028205','59267056501','80777028302','59267060901',
                   '5967658015','8077727399','8077727398','8077727399',
                   '8077727398','8077710099','8077727599','8077727999',
                   '8077728299','8077728399','5926710002','5926710003',
                   '5926710253','5926710254','0069202510','0069202525',
                   '5926710554','5926700784','5926703042','5926714042',
                   '5926705652','5926706092','8063110010','4928161820',
                   '0069100002','00069100003','8077710098','8077727799',
                   '8077728099','5967658005','8077727310','8077727315',
                   '8077727310','8077727315','8077710011','8077727505',
                   '8077727905','8077728205','8077728302','5926710001',
                   '5926710251','0069202501','5926710551','5926700781',
                   '5926703041','5926714041','5926705651','5926706091',
                   '8063110001','4928161878','0069100001','8077710015',
                   '8077727705','8077728005',]))
print(len(ndc_codes))

# Extract COVID-19 vaccination records

## from `immunization`

print(immunization_cancer.columns)
print(immunization_cancer.VX_CODE_TYPE.value_counts())

immunization_cancer_vax = extract_COVIDvax_record(df=immunization_cancer,\
                             extract_from_col="VX_CODE",\
                             code_type_col="VX_CODE_TYPE",\
                             cpt_codes=cpt_codes,\
                             rxnorm_codes=rxnorm_codes,\
                             ndc_codes=ndc_codes,\
                             cvx_codes=cvx_codes)

immunization_cancer_vax.isnull().sum()

## from `procedure`

print(procedure_cancer.columns)
print(procedure_cancer.PX_TYPE.value_counts())

procedure_cancer_vax = extract_COVIDvax_record(df=procedure_cancer,\
                             extract_from_col="PX",\
                             code_type_col="PX_TYPE",\
                             cpt_codes=cpt_codes,\
                             rxnorm_codes=rxnorm_codes,\
                             ndc_codes=ndc_codes,\
                             cvx_codes=cvx_codes)

procedure_cancer_vax.isnull().sum()

## from `prescribing`

print(prescribing_cancer.columns)

prescribing_cancer['CODE_TYPE']='RX'

prescribing_cancer['RXNORM_CUI'] = prescribing_cancer['RXNORM_CUI'].fillna('-9999999999').astype(int).astype(str)

prescribing_cancer_vax = extract_COVIDvax_record(df=prescribing_cancer,\
                             extract_from_col="RXNORM_CUI",\
                             code_type_col="CODE_TYPE",\
                             cpt_codes=cpt_codes,\
                             rxnorm_codes=rxnorm_codes,\
                             ndc_codes=ndc_codes,\
                             cvx_codes=cvx_codes)

prescribing_cancer_vax.isnull().sum()

## from `dispensing`

print(dispensing_cancer.columns)

dispensing_cancer_vax.isnull().sum()

dispensing_cancer['CODE_TYPE']='ND'

dispensing_cancer['NDC'] = dispensing_cancer['NDC'].fillna('-9999999999').astype(str)

dispensing_cancer_vax = extract_COVIDvax_record(df=dispensing_cancer,\
                             extract_from_col="NDC",\
                             code_type_col="CODE_TYPE",\
                             cpt_codes=cpt_codes,\
                             rxnorm_codes=rxnorm_codes,\
                             ndc_codes=ndc_codes,\
                             cvx_codes=cvx_codes)

dispensing_cancer_vax.isnull().sum()

# Assemble the data

print(len(immunization_cancer_vax['IMMUNIZATIONID']))
print(len(set(immunization_cancer_vax['IMMUNIZATIONID'])))
print(len(procedure_cancer_vax['PROCEDURESID']))
print(len(set(procedure_cancer_vax['PROCEDURESID'])))
print(len(prescribing_cancer_vax['PRESCRIBINGID']))
print(len(set(prescribing_cancer_vax['PRESCRIBINGID'])))
print(len(dispensing_cancer_vax['DISPENSINGID']))
print(len(set(dispensing_cancer_vax['DISPENSINGID'])))

immunization_cancer_vax = immunization_cancer_vax.rename(columns={"VX_ADMIN_DATE":"VX_DATE"})
prescribing_cancer_vax = prescribing_cancer_vax.rename(columns={"RX_ORDER_DATE":"VX_DATE"})
procedure_cancer_vax = procedure_cancer_vax.rename(columns={"PX_DATE":"VX_DATE"})
dispensing_cancer_vax = dispensing_cancer_vax.rename(columns={"DISPENSE_DATE":"VX_DATE"})

immunization_cancer_vax = immunization_cancer_vax.rename(columns={"IMMUNIZATIONID":"VX_ID"})
immunization_cancer_vax['VX_ID'] = "IMMUNIZATION"+immunization_cancer_vax['VX_ID']
immunization_cancer_vax['DOMAIN'] = "IMMUNIZATION"

prescribing_cancer_vax = prescribing_cancer_vax.rename(columns={"PRESCRIBINGID":"VX_ID"})
prescribing_cancer_vax['VX_ID'] = "PRESCRIBING"+prescribing_cancer_vax['VX_ID']
prescribing_cancer_vax['DOMAIN'] = "PRESCRIBING"

procedure_cancer_vax = procedure_cancer_vax.rename(columns={"PROCEDURESID":"VX_ID"})
procedure_cancer_vax['VX_ID'] = "PROCEDURE"+procedure_cancer_vax['VX_ID']
procedure_cancer_vax['DOMAIN'] = "PROCEDURE"

dispensing_cancer_vax = dispensing_cancer_vax.rename(columns={"DISPENSINGID":"VX_ID"})
dispensing_cancer_vax['VX_ID'] = "DISPENSING"+dispensing_cancer_vax['VX_ID']
dispensing_cancer_vax['DOMAIN'] = "DISPENSING"

vax_cancer = pd.concat([immunization_cancer_vax[['ID','VX_ID','VX_CODE_TYPE','VX_CODE','VX_DATE','DOMAIN']],
                        prescribing_cancer_vax[['ID','VX_ID','VX_CODE_TYPE','VX_CODE','VX_DATE','DOMAIN']],
                        procedure_cancer_vax[['ID','VX_ID','VX_CODE_TYPE','VX_CODE','VX_DATE','DOMAIN']],
                        dispensing_cancer_vax[['ID','VX_ID','VX_CODE_TYPE','VX_CODE','VX_DATE','DOMAIN']]
                       ])

vax_cancer['VX_ID_RAW'] = vax_cancer['VX_ID']
vax_cancer['VX_ID'] = vax_cancer['ID']+vax_cancer['VX_CODE']+vax_cancer['VX_DATE']

print(vax_cancer.shape)

#drop duplicate information
vax_cancer = vax_cancer.drop_duplicates('VX_ID',ignore_index=True)

print(vax_cancer.shape)

pickle.dump(vax_cancer, open('/data/datasets/changyuyin/2_COVIDwithCancer/vax_cancer.pkl','wb'))

pickle.dump(immunization_cancer_vax, open('/data/datasets/changyuyin/2_COVIDwithCancer/immunization_cancer_vax.pkl','wb'))
pickle.dump(procedure_cancer_vax, open('/data/datasets/changyuyin/2_COVIDwithCancer/procedure_cancer_vax.pkl','wb'))
pickle.dump(prescribing_cancer_vax, open('/data/datasets/changyuyin/2_COVIDwithCancer/prescribing_cancer_vax.pkl','wb'))
pickle.dump(dispensing_cancer_vax, open('/data/datasets/changyuyin/2_COVIDwithCancer/dispensing_cancer_vax.pkl','wb'))
