==> 6_EDA_subtypes&vax.ipynb <==
#package loading 
import sklearn
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from lifelines import CoxPHFitter
from dateutil.relativedelta import relativedelta
from scipy.stats import percentileofscore
import umap
from tqdm import tqdm
import cudf
import pickle
import sys

from utility import *
from preprocess import *

import warnings
warnings.filterwarnings("ignore")

death_cancer=pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/death_cancer.pkl','rb'))

vax_cancer=pickle.load( open('/data/datasets/changyuyin/2_COVIDwithCancer/vax_cancer.pkl','rb'))

all_cancer_pid = pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/pid_dx_cancer_subtypes.pkl','rb'))

# Patient number by cancer subtypesdeath_cancer

## Aggregation by vital status

subtype_summary,all_cancer_pid = cancer_subtype_summary(all_cancer_pid,death_cancer)

len(set(all_cancer_pid.loc[all_cancer_pid['IS_ALIVE']=="ALIVE"].ID))

subtype_summary

subtype_summary.to_csv('/home/changyu.yin/2.COVIDwithCancer/99_outputs/EDA_subtype_summary.csv')

## Co-occurrence matrix

cooccurrence_matrix = cancer_subtype_cooccurrence(all_cancer_pid)

cooccurrence_matrix

cooccurrence_matrix.to_csv('/home/changyu.yin/2.COVIDwithCancer/99_outputs/EDA_subtype_cooccurrence_matrix.csv')

# Vaccination summary

vax_cancer[['DOMAIN','VX_CODE_TYPE']].value_counts()

all_cancer_pid_vax_agg, all_cancer_pid_vax, vax_num_agg = cancer_subtype_vax(all_cancer_pid,vax_cancer)

vax_num_agg

vax_num_agg['N_vax_record'].value_counts()

vax_cancer.loc[vax_cancer['ID']=='19Oct2022covidpasc00046495']

#abnormal value check
prescribing_cancer_vax = pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/prescribing_cancer_vax.pkl','rb'))
prescribing_cancer_vax.loc[prescribing_cancer_vax['ID']=='19Oct2022covidpasc00147983'].sort_values(['VX_CODE','VX_DATE'])

#abnormal value check
procedure_cancer_vax = pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/procedure_cancer_vax.pkl','rb'))
procedure_cancer_vax.loc[procedure_cancer_vax['ID']=='19Oct2022covidpasc00147983'].sort_values(['VX_CODE','VX_DATE'])

vax_num_agg['N_vax_record_cap'].value_counts()

vax_num_agg['N_vax'].value_counts()

all_cancer_pid_vax_agg

all_cancer_pid_vax_agg.to_csv('/home/changyu.yin/2.COVIDwithCancer/99_outputs/EDA_N_vax_by_subtype.csv')
