==> 99_lungcancerAFTERcovid.ipynb <==
#package loading 
import sklearn
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from dateutil.relativedelta import relativedelta
from scipy.stats import percentileofscore
from tqdm import tqdm
import cudf
import numba as nb
import os
import sys
import time
import pickle

import sklearn
import numba as nb
import functools as ft
import swifter

from utility import *
#del sys.modules["preprocess"]
from preprocess import * 

import warnings
warnings.filterwarnings("ignore")

#define function for calculating ages
def calc_age(born_d,index_d):
   born_d = pd.to_datetime(born_d)
   index_d = pd.to_datetime(index_d)
   return index_d.year - born_d.year - ((index_d.month, 
                                      index_d.day) < (born_d.month, 
                                                    born_d.day))

pid_dx_subtypes = pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/pid_dx_cancer_subtypes.pkl','rb'))

#pasc and covid inex dates
index_date = pd.read_csv('/data/datasets/Tianchen/2022_covid_pasc_1FL_data/20221020/covid_pasc_cohort.csv')

data_wd = "/data/datasets/Tianchen/2022_covid_pasc_1FL_data/20221020/covid_database"

# covid database
diagnosis_cov = pd.read_csv(data_wd+"/DIAGNOSIS.csv",header=0,skiprows=[1])
procedure_cov = pd.read_csv(data_wd+"/PROCEDURES.csv",header=0,skiprows=[1])
demographics_cov = pd.read_csv(data_wd+"/DEMOGRAPHIC.csv",header=0,skiprows=[1])

data_wd = "/data/datasets/Tianchen/2022_covid_pasc_1FL_data/20221020/main_database"

# general database
diagnosis_m = pd.read_csv(data_wd+"/DIAGNOSIS.csv",header=0,skiprows=[1])
procedure_m = pd.read_csv(data_wd+"/PROCEDURES.csv",header=0,skiprows=[1])
demographics_m = pd.read_csv(data_wd+"/DEMOGRAPHIC.csv",header=0,skiprows=[1])

# union them
diagnosis = pd.concat([diagnosis_m,diagnosis_cov])
print("complete diagnosis")
procedure = pd.concat([procedure_m,procedure_cov])
print("complete procedure")
demographics = pd.concat([demographics_m,demographics_cov])
print("complete demographics")

#release memory
del(diagnosis_m)
del(diagnosis_cov)
del(procedure_m)
del(procedure_cov)
del(demographics_m)
del(demographics_cov)

# check the number of lung cancer patients that get the cancer after covid infection

## extract pid of lung cancers

pid_lung = list(set(pid_dx_subtypes.loc[pid_dx_subtypes['subtype']=="Lung"].reset_index(drop=True).ID))

## matching demographics and index dates

#choose those with covid19 index date
index_date = index_date.loc[~index_date['index_date'].isna()]

index_date['index_date'] = pd.to_datetime(index_date['index_date'])
index_date['pasc_onset_date'] = pd.to_datetime(index_date['pasc_onset_date'])

#index_date = index_date.loc[(index_date['index_date']<index_date['pasc_onset_date'])|index_date['pasc_onset_date'].isna()]

# This is because there are two database to be merged, i.e. covid database and general databse. The medical records are not duplicates abut the patient demographics are.
print(demographics.shape)
demographics = demographics.drop_duplicates(['ID'])
print(demographics.shape)

#pair demographics and index date
demographics = pd.merge(demographics,index_date,left_on='ID',right_on='ID',how='inner')
print(demographics.shape)

demographics = demographics[['ID',
                             'BIRTH_DATE',
                             'SEX',
                             'HISPANIC',
                             'RACE',
                             'PAT_PREF_LANGUAGE_SPOKEN',
                             'ZIP_CODE',
                             'index_date']]

demographics['index_age'] = demographics.swifter.apply(lambda x: calc_age(x.BIRTH_DATE, x.index_date), axis=1)

demographics

## get the 1st diagnosis date

dx_codes_icd9_Lungcaner = tuple(['1622', '1623', '1624', '1625', '1628', '1629', '2312', 'V1011'])
dx_codes_icd10_Lungcaner = tuple(['C34'])

#diagnosis
pid_dx_lung_diag = extract_cancer_pid_time(df=diagnosis,\
                                     extract_from_col="DX",\
                                     code_type_col="DX_TYPE",\
                                     date_col = "DX_DATE",\
                                     codes_icd9=dx_codes_icd9_Lungcaner,\
                                     codes_icd10=dx_codes_icd10_Lungcaner,\
                                     subtype="lung")

#procedure
pid_dx_lung_procedure = extract_cancer_pid_time(df=procedure,\
                                          extract_from_col="PX",\
                                          code_type_col="PX_TYPE",\
                                          date_col = "PX_DATE",\
                                          codes_icd9=dx_codes_icd9_Lungcaner,\
                                          codes_icd10=dx_codes_icd10_Lungcaner,\
                                          subtype="lung")

pid_dx_lung = pd.concat([pid_dx_lung_diag,pid_dx_lung_procedure]).\
                drop_duplicates(ignore_index=True)

pid_dx_lung['DX_DATE'] = pd.to_datetime(pid_dx_lung['DX_DATE'])
pid_dx_lung['ADMIT_DATE'] = pd.to_datetime(pid_dx_lung['ADMIT_DATE'])

dx_lung_date = pid_dx_lung.groupby('ID').agg({'DX_DATE':'min',
                               'ADMIT_DATE':'min'}).reset_index()

dx_lung_date['INDEX_DATE_LUNG'] = dx_lung_date[['DX_DATE','ADMIT_DATE']].apply('min',axis=1)

print(dx_lung_date.shape)
dx_lung_date = dx_lung_date.merge(demographics,on='ID',how='inner')
print(dx_lung_date.shape)

print(dx_lung_date.shape)

## get those with lung cancer after covid

dx_lung_afterCOVID = dx_lung_date.loc[dx_lung_date['INDEX_DATE_LUNG'] > dx_lung_date['index_date']].reset_index(drop=True)

dx_lung_afterCOVID['RACE']=dx_lung_afterCOVID['RACE'].replace({'01':'American Indian or Alaska Native',
                            '02':'Asian',
                            '03':'Black or African American',
                            '04':'Native Hawaiian or Other Pacific Islander',
                            '05':'White',
                            '06':'Multiple race',
                            '07':'Refuse to answer',
                            'NI':'No information',
                            'UN':'Unknown',
                            'OT':'Other' })
dx_lung_afterCOVID['RACE'] = dx_lung_afterCOVID['RACE'].fillna('No information')
dx_lung_afterCOVID['RACE_Ethinic'] = np.where( (dx_lung_afterCOVID['HISPANIC']=='N') & (dx_lung_afterCOVID['RACE']=='White'),
                                              'NHW',
                                              np.where((dx_lung_afterCOVID['HISPANIC']=='N') & (dx_lung_afterCOVID['RACE']=='Black or African American'),
                                                      'NHB',
                                                      np.where(dx_lung_afterCOVID['HISPANIC']=='Y',
                                                               'Hispanics',
                                                              np.where(dx_lung_afterCOVID['RACE'].isin(['Unknown','No information']),
                                                                       'Unknown',
                                                                       'Other')
                                                      )
                                             ))

#calculate age at diagnosis (index_date provided)
dx_lung_afterCOVID['index_age'] = dx_lung_afterCOVID.swifter.apply(lambda x: calc_age(x.BIRTH_DATE, x.index_date), axis=1)
#calc group
dx_lung_afterCOVID['age_band'] = pd.cut(dx_lung_afterCOVID.index_age, \
                                    [-999,18,65,75,999],\
                                    labels = ['[0,18)','[18,64)','[65,75)','>=75'],\
                                    include_lowest=True,right=False)

dx_lung_afterCOVID['age_band2'] = pd.cut(dx_lung_afterCOVID.index_age, \
                                    [-999,20,30,40,45,50,60,70,80,999],\
                                    labels = ['[0,20)','[20,30)','[30,40)','[40,45)','[45,50)','[50,60)','[60,70)','[70,80)','80+'],\
                                    include_lowest=True,right=False)

dx_lung_afterCOVID

# CCI=0

#output to r
# covid database
data_wd = "/data/datasets/Tianchen/2022_covid_pasc_1FL_data/20221020/covid_database"
diagnosis_cov = pd.read_csv(data_wd+"/DIAGNOSIS.csv",header=0,skiprows=[1])
procedure_cov = pd.read_csv(data_wd+"/PROCEDURES.csv",header=0,skiprows=[1])

data_wd = "/data/datasets/Tianchen/2022_covid_pasc_1FL_data/20221020/main_database"
# general database
diagnosis_m = pd.read_csv(data_wd+"/DIAGNOSIS.csv",header=0,skiprows=[1])
procedure_m = pd.read_csv(data_wd+"/PROCEDURES.csv",header=0,skiprows=[1])

# union them
diagnosis = pd.concat([diagnosis_m,diagnosis_cov])
print("complete diagnosis")
procedure = pd.concat([procedure_m,procedure_cov])
print("complete procedure")

#release memory
del(diagnosis_m)
del(diagnosis_cov)
del(procedure_m)
del(procedure_cov)

diagnosis_uniq = diagnosis[['ID','DX','DX_TYPE']].drop_duplicates(ignore_index=True)
procedure_uniq = procedure[['ID','PX','PX_TYPE']].drop_duplicates(ignore_index=True)

diagnosis_uniq.to_csv('/data/datasets/changyuyin/2_COVIDwithCancer/diagnosis_uniq.csv',index=False)
procedure_uniq.to_csv('/data/datasets/changyuyin/2_COVIDwithCancer/procedure_uniq.csv',index=False)

diagnosis_uniq.shape,procedure_uniq.shape

from comorbidipy import comorbidity

diagnosis_uniq = pd.read_csv('/data/datasets/changyuyin/2_COVIDwithCancer/diagnosis_uniq.csv')

procedure_uniq = pd.read_csv('/data/datasets/changyuyin/2_COVIDwithCancer/procedure_uniq.csv')

procedure_uniq = procedure_uniq[procedure_uniq['PX_TYPE'].isin(['09','10',10,9,'9'])]
procedure_uniq.columns = ['ID','DX','DX_TYPE']

diagnosis_uniq = diagnosis_uniq[diagnosis_uniq['DX_TYPE'].isin(['09','10',10,9,'9'])]

df = pd.concat([diagnosis_uniq,procedure_uniq]).drop_duplicates(ignore_index=True)

df['DX'] = df['DX'].str.replace('[^a-zA-Z0-9]', '')

df = df.merge(demographics[['ID','index_age']],how='left',on='ID')

## calculate scores

charlson9 = comorbidity(df, 
                         "ID", 
                         code = "DX", 
                         age=None,
                         icd='icd9',
                         score = "charlson", 
                         variant="quan",
                         weighting="quan",
                         assign0 = False)

charlson10 = comorbidity(df, 
                         "ID", 
                         code = "DX", 
                         age=None,
                         icd='icd10',
                         score = "charlson", 
                         variant="quan",
                         weighting="quan",
                         assign0 = False)

charlson_all = pd.concat([charlson9,charlson10]).groupby('ID').sum().reset_index()

charlson_all = demographics[['ID']].merge(charlson_all,how='left',on='ID').fillna(0)

# Non-cancer covid patient

all_cancer_pid = pickle.load(open('/data/datasets/changyuyin/2_COVIDwithCancer/pid_dx_cancer_subtypes.pkl','rb'))

pt_noncancer = demographics[~demographics['ID'].isin(all_cancer_pid.ID)]

pt_noncancer['RACE']=pt_noncancer['RACE'].replace({'01':'American Indian or Alaska Native',
                            '02':'Asian',
                            '03':'Black or African American',
                            '04':'Native Hawaiian or Other Pacific Islander',
                            '05':'White',
                            '06':'Multiple race',
                            '07':'Refuse to answer',
                            'NI':'No information',
                            'UN':'Unknown',
                            'OT':'Other' })
pt_noncancer['RACE'] = pt_noncancer['RACE'].fillna('No information')
pt_noncancer['RACE_Ethinic'] = np.where( (pt_noncancer['HISPANIC']=='N') & (pt_noncancer['RACE']=='White'),
                                              'NHW',
                                              np.where((pt_noncancer['HISPANIC']=='N') & (pt_noncancer['RACE']=='Black or African American'),
                                                      'NHB',
                                                      np.where(pt_noncancer['HISPANIC']=='Y',
                                                               'Hispanics',
                                                              np.where(pt_noncancer['RACE'].isin(['Unknown','No information']),
                                                                       'Unknown',
                                                                       'Other')
                                                      )
                                             ))

#calculate age at diagnosis (index_date provided)
pt_noncancer['index_age'] = pt_noncancer.swifter.apply(lambda x: calc_age(x.BIRTH_DATE, x.index_date), axis=1)
#calc group
pt_noncancer['age_band'] = pd.cut(pt_noncancer.index_age, \
                                    [-999,18,65,75,999],\
                                    labels = ['[0,18)','[18,64)','[65,75)','>=75'],\
                                    include_lowest=True,right=False)

pt_noncancer['age_band2'] = pd.cut(pt_noncancer.index_age, \
                                     [-999,20,30,40,45,50,60,70,80,999],\
                                    labels = ['[0,20)','[20,30)','[30,40)','[40,45)','[45,50)','[50,60)','[60,70)','[70,80)','80+'],\
                                    include_lowest=True,right=False)

pt_noncancer['age_band2'].value_counts()

pt_noncancer = pt_noncancer.merge(charlson_all[['ID','comorbidity_score']],how='left',on='ID')

pt_noncancer_healthy = pt_noncancer[pt_noncancer['comorbidity_score']==0]

pt_noncancer_healthy['age_band2'].value_counts()

dx_lung_afterCOVID['age_band2'].value_counts()

len(set(dx_lung_afterCOVID.ID))

# summary

def demog_summary(demographics):
    demogrpahics_tmp = demographics[['ID','SEX','RACE_Ethinic','age_band']].drop_duplicates(ignore_index=True)
    print(demographics.SEX.value_counts())
    print(demographics.RACE_Ethinic.value_counts())
    print(demographics.age_band.value_counts())

demog_summary(pt_noncancer)

demog_summary(dx_lung_afterCOVID)
